+++ ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 ### [4.3.0] - 2023-08-12
 
+-   Add Test, change Layout in Pixel converter
+-   Add Test, change layout in JsonToTypescript Converter
+    \_ Fix Some CSS Changes in Page Header
+
+### [4.3.0] - 2023-08-12
+
 -   Refactor ImageGeneratorFromColors
 -   Refactor ColorPicker